backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
psycopg[binary,pool]==3.2.10
redis==5.0.8
python-dotenv==1.0.1